pagination and selection refactor
